--- app/src/main/java/com/example/sokol/mcprices/ProductsDownLoader.java
@@ -1,10 +1,0 @@
-package com.example.sokol.mcprices;
-
-/**
- * Created by sokol on 08.03.2017.
- */
-
-interface ProductsDownLoader {
-
-    void loadOrDownload(boolean isOutOfDate);
-}
