--- app/src/main/java/com/example/sokol/mcprices/ProductsLoader.java
@@ -1,11 +1,0 @@
-package com.example.sokol.mcprices;
-
-/**
- * Created by sokol on 08.03.2017.
- */
-
-interface ProductsLoader {
-    void load();
-
-    void loadError();
-}
